requirements.txt
<<<<<<< HEAD
scikit-optimize==0.7
scikit-learn==0.21.3
numpy==1.14.5
pandas==0.23.4


=======
numpy>=1.15.4
pandas>=0.24.1
scikit-learn>=0.21.3
scikit-optimize>=0.5.2
pyyaml
>>>>>>> dev
